pyproject.toml
[tool.poetry]
name = "maio-test"
version = "0.1.0"
description = ""
authors = ["Guillaume Alleon <guillaume.alleon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11, <3.12"
meross-iot = "^0.4.7.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
